apps/backend/requirements.txt
fastapi
uvicorn
python-multipart
supabase
langgraph
langchain-google-vertexai
google-cloud-texttospeech
requests
pydantic
python-dotenv
google-generativeai
